+++ node/resolvers/__mocks__/shippping.ts
@@ -1,33 +1,61 @@
 import { ADDRESS_TYPES, DELIVERY, PICKUP_IN_STORE } from '../../constants'
+import { EMPTY_ORDER_FORM } from './orderForm'
+
+const SLA = {
+  deliveryIds: [],
+  deliveryWindow: null,
+  listPrice: 0,
+  lockTTL: null,
+  pickupDistance: 0,
+  pickupPointId: null,
+  pickupStoreInfo: {
+    additionalInfo: null,
+    address: null,
+    dockId: null,
+    friendlyName: null,
+    isPickupStore: false,
+  },
+  polygonName: null,
+  shippingEstimateDate: null,
+  tax: 0,
+}
 
 const pickupSLA = {
+  ...SLA,
   availableDeliveryWindows: [],
   deliveryChannel: PICKUP_IN_STORE,
   id: 'pickup-SLA',
+  name: 'pickup-SLA',
   price: 100,
   shippingEstimate: '1d',
 }
 
 const deliverySLA = {
+  ...SLA,
   availableDeliveryWindows: [],
   deliveryChannel: DELIVERY,
   id: 'delivery-SLA',
+  name: 'delivery-SLA',
   price: 100,
   shippingEstimate: '1db',
 }
 
 const deliverySLAExpress = {
+  ...SLA,
   availableDeliveryWindows: [],
   deliveryChannel: DELIVERY,
   id: 'delivery-SLA-Express',
+  name: 'delivery-SLA-Express',
   price: 600,
   shippingEstimate: '1h',
 }
 
 const scheduledDeliverySLA = {
+  ...SLA,
   availableDeliveryWindows: [{}],
   deliveryChannel: DELIVERY,
   id: 'scheduled-delivery-SLA',
+  name: 'scheduled-delivery-SLA',
   price: 100,
   shippingEstimate: '1bd',
 }
@@ -35,15 +63,29 @@ const scheduledDeliverySLA = {
 export const deliveryAddress = {
   addressId: 'test',
   addressType: ADDRESS_TYPES.RESIDENTIAL,
+  city: '',
+  complement: '',
   country: 'BRA',
+  geoCoordinates: [],
+  neighborhood: '',
+  number: '',
   postalCode: '222071060',
+  receiverName: '',
+  reference: null,
+  state: '',
+  street: '',
 }
 
 export const ORDER_FORM_WITH_PICKUPS = {
+  ...EMPTY_ORDER_FORM,
   shippingData: {
-    availableAddress: [deliveryAddress],
+    address: null,
+    availableAddresses: [deliveryAddress],
     logisticsInfo: [
       {
+        addressId: deliveryAddress.addressId,
+        deliveryChannels: [{ id: DELIVERY }, { id: PICKUP_IN_STORE }],
+        itemId: 'testId',
         itemIndex: 0,
         selectedDeliveryChannel: DELIVERY,
         selectedSla: deliverySLA.id,
@@ -51,15 +93,21 @@ export const ORDER_FORM_WITH_PICKUPS = {
         slas: [deliverySLA, pickupSLA],
       },
     ],
+    pickupPoints: [],
     selectedAddresses: [deliveryAddress],
   },
 }
 
 export const ORDER_FORM_WITH_SCHEDULED_DELIVERY = {
+  ...EMPTY_ORDER_FORM,
   shippingData: {
-    availableAddress: [deliveryAddress],
+    address: null,
+    availableAddresses: [deliveryAddress],
     logisticsInfo: [
       {
+        addressId: deliveryAddress.addressId,
+        deliveryChannels: [{ id: DELIVERY }, { id: PICKUP_IN_STORE }],
+        itemId: 'testId',
         itemIndex: 0,
         selectedDeliveryChannel: DELIVERY,
         selectedSla: deliverySLA.id,
@@ -67,15 +115,21 @@ export const ORDER_FORM_WITH_SCHEDULED_DELIVERY = {
         slas: [deliverySLA, scheduledDeliverySLA],
       },
     ],
+    pickupPoints: [],
     selectedAddresses: [deliveryAddress],
   },
 }
 
 export const ORDER_FORM_WITH_SCHEDULED_DELIVERY_AND_PICKUPS = {
+  ...EMPTY_ORDER_FORM,
   shippingData: {
-    availableAddress: [deliveryAddress],
+    address: null,
+    availableAddresses: [deliveryAddress],
     logisticsInfo: [
       {
+        addressId: deliveryAddress.addressId,
+        deliveryChannels: [{ id: DELIVERY }, { id: PICKUP_IN_STORE }],
+        itemId: 'testId',
         itemIndex: 0,
         selectedDeliveryChannel: DELIVERY,
         selectedSla: deliverySLA.id,
@@ -83,15 +137,21 @@ export const ORDER_FORM_WITH_SCHEDULED_DELIVERY_AND_PICKUPS = {
         slas: [deliverySLA, scheduledDeliverySLA, pickupSLA],
       },
     ],
+    pickupPoints: [],
     selectedAddresses: [deliveryAddress],
   },
 }
 
 export const ORDER_FORM_WITH_DIFFERENT_SLAS_BETWEEN_LOGISTICS_INFO = {
+  ...EMPTY_ORDER_FORM,
   shippingData: {
-    availableAddress: [deliveryAddress],
+    address: null,
+    availableAddresses: [deliveryAddress],
     logisticsInfo: [
       {
+        addressId: deliveryAddress.addressId,
+        deliveryChannels: [{ id: DELIVERY }, { id: PICKUP_IN_STORE }],
+        itemId: 'testId',
         itemIndex: 0,
         selectedDeliveryChannel: DELIVERY,
         selectedSla: deliverySLA.id,
@@ -99,6 +159,9 @@ export const ORDER_FORM_WITH_DIFFERENT_SLAS_BETWEEN_LOGISTICS_INFO = {
         slas: [deliverySLA],
       },
       {
+        addressId: deliveryAddress.addressId,
+        deliveryChannels: [{ id: DELIVERY }, { id: PICKUP_IN_STORE }],
+        itemId: 'testId',
         itemIndex: 1,
         selectedDeliveryChannel: DELIVERY,
         selectedSla: deliverySLA.id,
@@ -106,15 +169,21 @@ export const ORDER_FORM_WITH_DIFFERENT_SLAS_BETWEEN_LOGISTICS_INFO = {
         slas: [deliverySLA, deliverySLAExpress],
       },
     ],
+    pickupPoints: [],
     selectedAddresses: [deliveryAddress],
   },
 }
 
 export const ORDER_FORM_WITH_EMPTY_LOGISTICS_INFO = {
+  ...EMPTY_ORDER_FORM,
   shippingData: {
-    availableAddress: [deliveryAddress],
+    address: null,
+    availableAddresses: [deliveryAddress],
     logisticsInfo: [
       {
+        addressId: deliveryAddress.addressId,
+        deliveryChannels: [{ id: DELIVERY }, { id: PICKUP_IN_STORE }],
+        itemId: 'testId',
         itemIndex: 0,
         selectedDeliveryChannel: DELIVERY,
         selectedSla: null,
@@ -122,6 +191,9 @@ export const ORDER_FORM_WITH_EMPTY_LOGISTICS_INFO = {
         slas: [],
       },
       {
+        addressId: deliveryAddress.addressId,
+        deliveryChannels: [{ id: DELIVERY }, { id: PICKUP_IN_STORE }],
+        itemId: 'testId',
         itemIndex: 1,
         selectedDeliveryChannel: DELIVERY,
         selectedSla: null,
@@ -129,8 +201,11 @@ export const ORDER_FORM_WITH_EMPTY_LOGISTICS_INFO = {
         slas: [],
       },
     ],
+    pickupPoints: [],
     selectedAddresses: [deliveryAddress],
   },
 }
 
-export const ORDER_FORM_WITH_EMPTY_SHIPPING_DATA = {}
+export const ORDER_FORM_WITH_EMPTY_SHIPPING_DATA = {
+  ...EMPTY_ORDER_FORM,
+}