+++ node/resolvers/shipping.ts
@@ -1,18 +1,22 @@
 import flatten from 'lodash/flatten'
 import uniq from 'lodash/uniq'
+import { DELIVERY, ADDRESS_TYPES } from '../constants'
 import { getNewOrderForm } from './orderForm'
 
 const getShippingData = (
   address: AddressInput,
-  logisticsInfo: LogisticsInfo[]
+  logisticsInfo: LogisticsInfo[] | null
 ) => {
   const selectedAddresses = [address]
   const hasGeocoordinates =
     address && address.geoCoordinates && address.geoCoordinates.length > 0
-  const logisticsInfoWithAddress = logisticsInfo.map((li: LogisticsInfo) => ({
-    ...li,
-    addressId: address.addressId,
-  }))
+  const logisticsInfoWithAddress =
+    (logisticsInfo &&
+      logisticsInfo.map((li: LogisticsInfo) => ({
+        ...li,
+        addressId: address.addressId,
+      }))) ||
+    []
 
   const requestPayload = {
     logisticsInfo: logisticsInfoWithAddress,
@@ -31,8 +35,20 @@ export const getShippingInfo = (orderForm: CheckoutOrderForm) => {
     flatten(logisticsInfo ? logisticsInfo.map(item => item.shipsTo) : [])
   )
 
-  const selectedAddresses =
-    (orderForm.shippingData && orderForm.shippingData.selectedAddresses) || []
+  const availableAddresses =
+    (orderForm.shippingData && orderForm.shippingData.availableAddresses) || []
+
+  const selectedAddress =
+    (orderForm.shippingData &&
+      orderForm.shippingData.selectedAddresses.find(address => {
+        const isCommercial = address.addressType === ADDRESS_TYPES.COMMERCIAL
+        const isResidential = address.addressType === ADDRESS_TYPES.RESIDENTIAL
+        const iGiftRegistry =
+          address.addressType === ADDRESS_TYPES.GIFT_REGISTRY
+
+        return isCommercial || isResidential || iGiftRegistry
+      })) ||
+    null
 
   const deliveryOptions = uniq(
     flatten(logisticsInfo ? logisticsInfo.map(item => item.slas) : [])
@@ -44,13 +60,14 @@ export const getShippingInfo = (orderForm: CheckoutOrderForm) => {
   const filteredDeliveryOptions = deliveryOptions.filter(
     (deliveryOption: SLA) => {
       const deliveryOptionIsInEveryLogisticsInfo =
+        logisticsInfo &&
         logisticsInfo.length > 0 &&
         logisticsInfo.every((li: LogisticsInfo) =>
           li.slas.some(sla => sla.id === deliveryOption.id)
         )
 
       return (
-        deliveryOption.deliveryChannel === 'delivery' &&
+        deliveryOption.deliveryChannel === DELIVERY &&
         deliveryOption.availableDeliveryWindows.length === 0 &&
         deliveryOptionIsInEveryLogisticsInfo
       )
@@ -60,14 +77,19 @@ export const getShippingInfo = (orderForm: CheckoutOrderForm) => {
   const updatedDeliveryOptions = filteredDeliveryOptions.map(sla => {
     let price = 0
 
-    const isSelected = logisticsInfo.some(li => li.selectedSla === sla.id)
-    logisticsInfo.forEach((li: LogisticsInfo) => {
-      const currentSla = li.slas.find(liSla => liSla.id === sla.id)
+    const isSelected =
+      (logisticsInfo && logisticsInfo.some(li => li.selectedSla === sla.id)) ||
+      false
 
-      if (currentSla) {
-        price += currentSla.price
-      }
-    })
+    if (logisticsInfo) {
+      logisticsInfo.forEach((li: LogisticsInfo) => {
+        const currentSla = li.slas.find(liSla => liSla.id === sla.id)
+
+        if (currentSla) {
+          price += currentSla.price
+        }
+      })
+    }
 
     return {
       estimate: sla.shippingEstimate,
@@ -78,9 +100,10 @@ export const getShippingInfo = (orderForm: CheckoutOrderForm) => {
   })
 
   return {
+    availableAddresses,
     countries,
     deliveryOptions: updatedDeliveryOptions,
-    selectedAddresses,
+    selectedAddress,
   }
 }