+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- SLA filtering to return only Delivery SLAs
+
+### Changed
+
+- Shipping to return array of selected addresses
+
 ## [0.11.0] - 2019-09-23
 
 ### Added